Pass tolerance options to numpy array comparison
- Widen `Options.absolute_tolerance` type from `int` to `float` so it maps cleanly to numpy's `atol` parameter
- Add three test cases covering loose rtol, tight rtol, and loose atol to verify the tolerances are respected during numpy array comparison
- Pass `o.relative_tolerance` and `o.absolute_tolerance` as `rtol`/`atol` to `array_compare()` in `function_return_values_match_reference.py`

+++ src/generic_grader/utils/options.py
@@ -65,7 +65,7 @@ class Options:
     # Stats
     expected_distribution: dict = {0: 0}
     relative_tolerance: float = 1e-7
-    absolute_tolerance: int = 0
+    absolute_tolerance: float = 0.0
 
     # Image
     mode: str = "exactly"
